test: fixed profile tests

+++ server/src/test/kotlin/it/polito/wa2/g19/server/integration/profile/ProfileTest.kt
@@ -3,25 +3,15 @@ package it.polito.wa2.g19.server.integration.profile
 import com.nimbusds.jose.shaded.gson.Gson
 import dasniko.testcontainers.keycloak.KeycloakContainer
 import it.polito.wa2.g19.server.Util
-import it.polito.wa2.g19.server.integration.chat.ChatTest
-import it.polito.wa2.g19.server.main
-import it.polito.wa2.g19.server.products.Product
-import it.polito.wa2.g19.server.products.ProductRepository
+
 import it.polito.wa2.g19.server.profiles.DuplicateEmailException
+import it.polito.wa2.g19.server.profiles.KeycloakException
 import it.polito.wa2.g19.server.profiles.LoginDTO
 import it.polito.wa2.g19.server.profiles.customers.CredentialCustomerDTO
-import it.polito.wa2.g19.server.profiles.customers.Customer
 import it.polito.wa2.g19.server.profiles.customers.CustomerDTO
 import it.polito.wa2.g19.server.profiles.customers.CustomerRepository
 import it.polito.wa2.g19.server.profiles.staff.*
-import it.polito.wa2.g19.server.ticketing.attachments.AttachmentRepository
-import it.polito.wa2.g19.server.ticketing.chat.ChatMessageOutDTO
-import it.polito.wa2.g19.server.ticketing.chat.ChatMessageRepository
-import it.polito.wa2.g19.server.ticketing.statuses.PriorityLevelEnum
-import it.polito.wa2.g19.server.ticketing.statuses.TicketStatusEnum
-import it.polito.wa2.g19.server.ticketing.statuses.TicketStatusRepository
-import it.polito.wa2.g19.server.ticketing.tickets.PriorityLevelRepository
-import it.polito.wa2.g19.server.ticketing.tickets.TicketRepository
+
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
@@ -34,11 +24,11 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.boot.test.web.client.exchange
 import org.springframework.boot.test.web.client.postForEntity
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
+import org.springframework.http.ProblemDetail
 import org.springframework.http.ResponseEntity
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
@@ -95,14 +85,9 @@ class ProfileTest {
     private var mainCustomerEmail = "newCustomer@test.it".lowercase()
     private var mainExpertEmail = "newExpert@test.it".lowercase()
 
-    private lateinit var customer: Customer
-    private lateinit var otherCustomer: Customer
-    private lateinit var product: Product
+
     private lateinit var manager: Manager
-    private lateinit var expert: Expert
-    private lateinit var otherExpert: Expert
-    private lateinit var customerToken: String
-    private lateinit var expertToken: String
+
     private lateinit var managerToken: String
 
     @Autowired
@@ -111,27 +96,11 @@ class ProfileTest {
     @Autowired
     lateinit var staffRepository: StaffRepository
 
-    @Autowired
-    lateinit var productRepository: ProductRepository
-
-    @Autowired
-    lateinit var ticketRepository: TicketRepository
-
-    @Autowired
-    lateinit var priorityLevelRepository: PriorityLevelRepository
-
-    @Autowired
-    lateinit var ticketStatusRepository: TicketStatusRepository
-
-    @Autowired
-    lateinit var chatRepository: ChatMessageRepository
-
-    @Autowired
-    lateinit var attachmentRepository: AttachmentRepository
 
 
     @BeforeEach
     fun populateDatabase() {
+        if(!postgres.isRunning) postgres.start()
         println("----populating database------")
         manager = Util.mockMainManager()
         manager.id = UUID.randomUUID()
@@ -143,12 +112,17 @@ class ProfileTest {
     @AfterEach
     fun destroyDatabase() {
         println("----destroying database------")
-
-        keycloak.realm(realmName).users().search(mainCustomerEmail).let {
-            if (it.isEmpty()) return@let null
-            val user = it.first()
-            keycloak.realm(realmName).users().delete(user.id)
+        for(u in listOf(mainCustomerEmail, mainExpertEmail)){
+            keycloak.realm(realmName).users().search(u).let {
+                if (it.isEmpty()) return@let null
+                val user = it.first()
+                keycloak.realm(realmName).users().delete(user.id)
+            }
         }
+        if (!postgres.isRunning){
+            return
+        }
+
         customerRepository.deleteAll()
         staffRepository.deleteAll()
         println("---------------------------------")
@@ -176,8 +150,8 @@ class ProfileTest {
     }
 
     fun mockExpertDTO(): StaffDTO{
-        return StaffDTO(mainExpertEmail, "newCustomerName",
-            "newCustomerSurname", StaffType.Expert, listOf("Hardware"))
+        return StaffDTO(mainExpertEmail, "newExpertName",
+            "newExpertSurname", StaffType.Expert, listOf("Hardware"))
 
     }
 
@@ -205,11 +179,13 @@ class ProfileTest {
         assert(ku.attributes["address"]!!.first() == newCustomer.customerDTO.address)
         assert(ku.username == newCustomer.customerDTO.email)
         assert(ku.firstName == newCustomer.customerDTO.name)
+        assert(keycloak.realm(realmName).users().searchByUsername(mainCustomerEmail, true).isNotEmpty() )
+
 
     }
 
     @Test
-    fun `signup and signin is successfful`(){
+    fun `signup and login is successful`(){
         val newCustomer = mockCredentialCustomerDTO()
         val request = HttpEntity(newCustomer)
         val response = restTemplate.postForEntity<Unit>("/API/signup", request, HttpMethod.POST)
@@ -217,8 +193,6 @@ class ProfileTest {
 
         val loginResponse: ResponseEntity<String> = restTemplate.exchange("/API/login", HttpMethod.POST, HttpEntity(LoginDTO(newCustomer.customerDTO.email, newCustomer.password)))
         assert(loginResponse.statusCode == HttpStatus.OK)
-
-
     }
 
     @Test
@@ -234,7 +208,7 @@ class ProfileTest {
     }
 
     @Test
-    fun `signup without password is unsuccesful`(){
+    fun `signup without password is unsuccessful`(){
         val newCustomer = mockCredentialCustomerDTO()
         val customerJson = hashMapOf<String, Any>(
             "customerDTO" to newCustomer.customerDTO
@@ -248,8 +222,10 @@ class ProfileTest {
 
 
 
+
+
     @Test
-    fun `create a expert is successfull`(){
+    fun `create a expert is successful`(){
         val newExpert = mockCredentialStaffDTO()
         val headers = HttpHeaders()
         headers.setBearerAuth(managerToken)
@@ -261,7 +237,7 @@ class ProfileTest {
     }
 
     @Test
-    fun `create a expert and signin as him is successfull`(){
+    fun `create a expert and login as him is successful`(){
         val newExpert = mockCredentialStaffDTO()
         val headers = HttpHeaders()
         headers.setBearerAuth(managerToken)
@@ -276,7 +252,7 @@ class ProfileTest {
 
 
     @Test
-    fun `create the same expert is unsuccessfull`(){
+    fun `create the same expert is unsuccessful`(){
         val newExpert = mockCredentialStaffDTO()
         val headers = HttpHeaders()
         headers.setBearerAuth(managerToken)
@@ -290,18 +266,22 @@ class ProfileTest {
     }
 
     @Test
-    fun `create expert without password is unsuccesful`(){
+    fun `create expert without password is unsuccessful`(){
         val newExpert = mockCredentialStaffDTO()
         val expertJson = hashMapOf<String, Any>(
             "staffDTO" to newExpert.staffDTO
         )
+        val headers = HttpHeaders()
+        headers.setBearerAuth(managerToken)
 
-        val request = HttpEntity(expertJson)
+        val request = HttpEntity(expertJson, headers)
         val response = restTemplate.postForEntity<Unit>("/API/signup", request, HttpMethod.POST)
         assert(response.statusCode == HttpStatus.BAD_REQUEST)
         assert(customerRepository.findByEmailIgnoreCase(newExpert.staffDTO.email) == null)
         assert(keycloak.realm(realmName).users().searchByUsername(mainExpertEmail, true).isEmpty() )
     }
+
+
 
 
 }
